requirements.txt
hypster[jupyter]

ipywidgets
ipykernel

haystack-ai
qdrant-haystack
fastembed
fastembed-haystack
jina-haystack

pypdf

transformers[torch,sentencepiece]
python-dotenv
jupyterlab